third_party/blink/web_tests/external/wpt/custom-elements/custom-element-registry/per-global-expected.txt

This is a testharness.js-based test.
[FAIL] Navigating from the initial about:blank must not replace window.customElements
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
[PASS] Discarding the browsing context must not change window.customElements
[PASS] document.open() must not replace window.customElements
Harness: the test ran to completion.
